requirements.txt
matplotlib==3.5.2
xlrd==1.1.0
pandas==0.22.0
numpy~=1.23.4
Pillow~=9.4.0
scikit-learn==0.19.1
opencv-python~=4.7.0.72